// apps/ed-tech.co/app/NX/types.d.ts
export type T_Config = {
    siteName: string;
    tenant: string;
    description: string;
    url: string;
    owner: {
        name: string;
        email: string;
    };
    images: {
        light: string;
        dark: string;
    };
    favicon: string;
    avatars: {
        light: string;
        dark: string;
    };
    cartridges: {
        nxadmin?: {
            enabled: boolean;
            layout?: string;
        },
        orders?: {
            enabled: boolean;
            frontPage: string;
            adminPage: string;
        },
        async?: {
            enabled: boolean;
        },
        tings?: {
            enabled: boolean;
        },
        paywall?: {
            enabled: boolean;
        },
        designSystem?: {
            themeSwitching: boolean;
            defaultTheme: string;
            themes: {
                [key: string]: {
                    mode: string;
                    primary: string;
                    secondary: string;
                    background: string;
                    paper: string;
                    text: string;
                };
            };
        };
    };
}
// ...
export interface I_NX {
    children: React.ReactNode;
    config: T_Config;
    frontmatter?: T_Frontmatter;
    flash?: string;
}
// ...
export type T_Theme = {
    mode: 'light' | 'dark';
    primary: string;
    secondary: string;
    background: string;
    paper: string;
    text: string;
    border: string;
};

export type T_Frontmatter = {
    title?: string;
    description?: string;
    slug?: string;
    tags?: string;
    icon?: string;
    order?: number;
    image?: string;
    layout?: string;
    hideInNav?: boolean | string;
};